requirements.txt
flask
pandas
scikit-learn
joblib
python-whois
requests